work in progress: refactor types and cleanup
