+++ Cargo.toml
@@ -2,13 +2,9 @@
 members = [
     "crates/ev-prover",
     "crates/ev-state-queries",
-    "crates/ev-state-types",
-    "crates/ev-storage-proofs",
-    "crates/ev-storage-proofs",
-    "crates/sp1/ev-exec-types",
+    "crates/ev-zkevm-types",
     "crates/sp1/ev-exec/program",
     "crates/sp1/ev-exec/script",
-    "crates/sp1/ev-hyperlane-types",
     "crates/sp1/ev-hyperlane/program",
     "crates/sp1/ev-hyperlane/script",
     "crates/sp1/ev-range-exec/program",
@@ -18,20 +14,14 @@ members = [
 resolver = "2"
 
 [workspace.dependencies]
-alloy-consensus = { version = "1.0.13" }
-alloy-genesis = { version = "1.0.13", default-features = false }
-alloy-primitives = { version = "1.2.0", default-features = false, features = [
-    "sha3-keccak",
-    "map-foldhash",
-] }
-alloy-provider = { version = "1.0.13", default-features = false, features = [
-    "reqwest",
-    "reqwest-rustls-tls",
-] }
-alloy-rlp = "0.3.10"
-alloy-rpc-types = { version = "1.0.27", default-features = false, features = [
-    "eth",
-] }
+alloy = { version = "1.0.32", default-features = false }
+alloy-genesis = { version = "1.0.32", default-features = false }
+alloy-trie = { version = "0.9.1", default-features = false }
+alloy-primitives = { version = "1.2.0", default-features = false }
+alloy-provider = { version = "1.0.32", default-features = false }
+alloy-rpc-types = { version = "1.0.32", default-features = false }
+alloy-consensus = { version = "1.0.32", default-features = false }
+alloy-rlp = { version = "0.3.12", default-features = false }
 anyhow = { version = "1.0", default-features = false }
 base64 = "0.22.1"
 bincode = "1.3.3"
@@ -40,11 +30,9 @@ celestia-rpc = "=0.13.0"
 celestia-types = "=0.15.0"
 ed25519-dalek = { version = "2.1.0", features = ["alloc", "serde"] }
 ethers = { version = "2.0", features = ["ws", "rustls"] }
-ev-hyperlane-types = { package = "ev-hyperlane-types", path = "crates/sp1/ev-hyperlane-types" }
 ev-state-queries = { path = "crates/ev-state-queries" }
-ev-state-types = { path = "crates/ev-state-types" }
-ev-storage-proofs = { path = "crates/ev-storage-proofs" }
 ev-types = { version = "0.0.1", default-features = false }
+ev-zkevm-types = { path = "crates/ev-zkevm-types" }
 hex = "0.4.3"
 jsonrpsee-core = "0.25.1"
 nmt-rs = "*"
@@ -78,7 +66,7 @@ tokio = { version = "1.45.0", features = [
 tracing = "0.1.40"
 
 [patch.crates-io]
-# Add support for precompiles: https://docs.succinct.xyz/docs/sp1/optimizing-programs/precompiles
+# Add support for precompiles: https://docs.succinct.xyz/docs/sp1/optimizing-programs/precompilesalloy-consensus = { version = "1.0.13" }
 sha2 = { git = "https://github.com/sp1-patches/RustCrypto-hashes", tag = "patch-sha2-0.10.9-sp1-4.0.0", package = "sha2" }
 bn = { git = "https://github.com/sp1-patches/bn", tag = "patch-0.6.0-sp1-5.0.0", package = "substrate-bn" }
 sha3 = { git = "https://github.com/sp1-patches/RustCrypto-hashes", tag = "patch-sha3-0.10.8-sp1-4.0.0" }
