+++ crates/ev-prover/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+alloy = { workspace = true }
 alloy-genesis = { workspace = true }
 alloy-primitives = { workspace = true }
 alloy-provider = { workspace = true }
@@ -16,12 +17,9 @@ celestia-rpc = { workspace = true }
 celestia-types = { workspace = true }
 clap = { version = "4", features = ["derive"] }
 dirs = "5"
-ev-exec-types = { path = "../sp1/ev-exec-types" }
-ev-hyperlane-types = { workspace = true }
 ev-state-queries = { workspace = true }
-ev-state-types = { workspace = true }
-ev-storage-proofs = { workspace = true, features = ["alloy"] }
 ev-types = { workspace = true, features = ["grpc"] }
+ev-zkevm-types = { workspace = true }
 hex = { workspace = true }
 jsonrpsee-core = { workspace = true }
 nmt-rs = { workspace = true }
