+++ crates/ev-prover/src/prover/programs/block.rs
@@ -15,8 +15,8 @@ use celestia_rpc::blob::BlobsAtHeight;
 use celestia_rpc::{client::Client, BlobClient, HeaderClient, ShareClient};
 use celestia_types::nmt::{Namespace, NamespaceProof};
 use celestia_types::Blob;
-use ev_exec_types::{BlockExecInput, BlockExecOutput};
 use ev_types::v1::SignedData;
+use ev_zkevm_types::programs::block::{BlockExecInput, BlockExecOutput};
 use jsonrpsee_core::client::Subscription;
 use prost::Message;
 use reth_chainspec::ChainSpec;
