+++ crates/ev-state-queries/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+alloy = { workspace = true }
 alloy-consensus = { workspace = true }
 alloy-primitives = { workspace = true }
 alloy-provider = { version = "1.0.2", default-features = false, features = [
@@ -15,8 +16,7 @@ alloy-sol-types = "1.3"
 anyhow = { workspace = true }
 async-trait = "0.1.88"
 dotenvy = "0.15.7"
-ev-state-types = { workspace = true }
-ev-storage-proofs = { workspace = true }
+ev-zkevm-types = { workspace = true }
 serde_json = { workspace = true }
 storage = { workspace = true }
 tokio = { workspace = true }
