--- crates/ev-state-types/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "ev-state-types"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-alloy-primitives = { workspace = true }
-alloy-sol-types = "1.3"
-anyhow = { workspace = true }
-hex = { workspace = true }
-serde = { workspace = true }
