--- crates/ev-storage-proofs/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "ev-storage-proofs"
-version = "0.1.0"
-edition = "2024"
-
-[features]
-alloy = ["dep:alloy", "dep:tokio"]
-
-[dependencies]
-alloy = { version = "1.0.27", features = ["full"], optional = true }
-alloy-primitives = { workspace = true }
-alloy-rlp = { version = "0.3.12", default-features = false }
-alloy-rpc-types = "1.0.27"
-alloy-trie = { version = "0.9.1", default-features = false, features = [
-    "ethereum",
-] }
-anyhow = "1"
-serde = { workspace = true }
-tokio = { version = "1.45.0", features = [
-    "sync",
-    "macros",
-    "io-util",
-    "rt",
-    "time",
-    "rt-multi-thread",
-], default-features = false, optional = true }
