+++ crates/ev-zkevm-types/Cargo.toml
@@ -1,12 +1,18 @@
 [package]
+name = "ev-zkevm-types"
 version = "0.1.0"
-name = "ev-exec-types"
-edition = "2021"
+edition = "2024"
 
 [dependencies]
 alloy-primitives = { workspace = true }
+alloy-sol-types = "1.3"
+anyhow = { workspace = true }
 bincode = { workspace = true }
 celestia-types = { workspace = true }
 hex = { workspace = true }
 rsp-client-executor = { workspace = true }
 serde = { workspace = true }
+sha3 = { workspace = true }
+alloy-rlp = { workspace = true }
+alloy-rpc-types = { workspace = true }
+alloy-trie = { workspace = true }
