+++ crates/ev-zkevm-types/src/programs/hyperlane/mod.rs
@@ -1,8 +1,7 @@
-use alloy_primitives::Keccak256;
-
-#[cfg(feature = "alloy")]
-pub mod client;
+pub mod tree;
 pub mod types;
+
+use sha3::{Digest, Keccak256};
 
 pub fn digest_keccak(bytes: &[u8]) -> [u8; 32] {
     let mut hasher = Keccak256::new();
