+++ crates/ev-zkevm-types/src/programs/hyperlane/types.rs
@@ -1,3 +1,10 @@
+use std::str::FromStr;
+
+use crate::hyperlane::HyperlaneMessage;
+use crate::programs::hyperlane::digest_keccak;
+use crate::programs::hyperlane::tree::{MerkleTree, ZERO_BYTES};
+use serde::{Deserialize, Serialize};
+
 use alloy_primitives::{
     Address, Bytes, FixedBytes, U256, Uint,
     hex::{FromHex, ToHexExt},
@@ -6,9 +13,6 @@ use alloy_rlp::encode;
 use alloy_rpc_types::EIP1186AccountProofResponse;
 use alloy_trie::{Nibbles, TrieAccount, proof::verify_proof};
 use anyhow::{Context, Result};
-use serde::{Deserialize, Serialize};
-
-use crate::digest_keccak;
 
 pub type MerkleProof = Vec<Vec<u8>>;
 pub type StoredValue = Vec<u8>;
@@ -241,5 +245,85 @@ impl HyperlaneBranchProofInputs {
             }
         }
         Ok(true)
+    }
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+/// Inputs for the hyperlane message circuit.
+pub struct HyperlaneMessageInputs {
+    pub state_root: String,
+    pub contract: String,
+    pub messages: Vec<HyperlaneMessage>,
+    pub branch_proof: HyperlaneBranchProofInputs,
+    pub snapshot: MerkleTree,
+}
+
+/// Implementation of the hyperlane message inputs.
+impl HyperlaneMessageInputs {
+    pub fn new(
+        state_root: String,
+        contract: String,
+        messages: Vec<HyperlaneMessage>,
+        branch_proof: HyperlaneBranchProofInputs,
+        snapshot: MerkleTree,
+    ) -> Self {
+        Self {
+            state_root,
+            contract,
+            messages,
+            branch_proof,
+            snapshot,
+        }
+    }
+
+    /// Verify the hyperlane message inputs against the branch proof and snapshot.
+    pub fn verify(&mut self) {
+        let message_ids: Vec<String> = self.messages.iter().map(|m| m.id()).collect();
+        for message_id in message_ids {
+            self.snapshot
+                .insert(message_id)
+                .expect("Failed to insert message id into snapshot");
+        }
+
+        // sanity check, we can't prove an empty hyperlane tree against state_root
+        if self
+            .snapshot
+            .branch
+            .iter()
+            .all(|_| self.snapshot.branch.iter().all(|b| b == ZERO_BYTES))
+        {
+            println!("Snapshot branch is empty (all zero bytes) before proof verification");
+        }
+
+        for idx in 0..HYPERLANE_MERKLE_TREE_KEYS.len() {
+            // The branch nodes of the snapshot after insert must match the branch nodes of the incremental
+            // tree on the EVM chain.
+            assert_eq!(self.snapshot.branch[idx], self.branch_proof.get_branch_node(idx));
+        }
+
+        let verified = self
+            .branch_proof
+            .verify(
+                &HYPERLANE_MERKLE_TREE_KEYS,
+                Address::from_str(&self.contract).unwrap(),
+                &self.state_root,
+            )
+            .expect("Failed to verify branch proof");
+        assert!(verified);
+    }
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct HyperlaneMessageOutputs {
+    pub state_root: [u8; 32],
+    pub message_ids: Vec<[u8; 32]>,
+}
+
+impl HyperlaneMessageOutputs {
+    pub fn new(state_root: [u8; 32], message_ids: Vec<[u8; 32]>) -> Self {
+        Self {
+            state_root,
+            message_ids,
+        }
     }
 }
