+++ crates/ev-zkevm-types/src/programs/mod.rs
@@ -1,0 +1,2 @@
+pub mod block;
+pub mod hyperlane;
