+++ crates/sp1/ev-exec/program/Cargo.toml
@@ -12,8 +12,8 @@ bincode = { workspace = true }
 bytes = { workspace = true }
 celestia-types = { workspace = true }
 ed25519-dalek = { workspace = true }
-ev-exec-types = { path = "../../ev-exec-types" }
 ev-types = { workspace = true, default-features = false }
+ev-zkevm-types = { workspace = true }
 hex = "0.4.3"
 nmt-rs = "*"
 prost = { workspace = true }
