+++ crates/sp1/ev-exec/program/src/main.rs
@@ -45,8 +45,8 @@ use bytes::Bytes;
 use celestia_types::nmt::NamespacedHash;
 use celestia_types::Blob;
 use ed25519_dalek::{Signature, Verifier, VerifyingKey};
-use ev_exec_types::{BlockExecInput, BlockExecOutput};
 use ev_types::v1::{Data, SignedData};
+use ev_zkevm_types::programs::block::{BlockExecInput, BlockExecOutput};
 use nmt_rs::NamespacedSha2Hasher;
 use prost::Message;
 use reth_primitives::TransactionSigned;
