+++ crates/sp1/ev-exec/script/src/bin/main.rs
@@ -27,7 +27,7 @@ use anyhow::Result;
 use celestia_types::nmt::{Namespace, NamespaceProof};
 use celestia_types::{Blob, DataAvailabilityHeader};
 use clap::Parser;
-use ev_exec_types::{BlockExecInput, BlockExecOutput};
+use ev_zkevm_types::programs::block::{BlockExecInput, BlockExecOutput};
 use hashbrown::HashMap;
 use rsp_client_executor::io::{EthClientExecutorInput, WitnessInput};
 use serde::{Deserialize, Serialize};
