--- crates/sp1/ev-hyperlane-types/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "ev-hyperlane-types"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-alloy-primitives = { workspace = true }
-anyhow = { workspace = true }
-ev-state-types = { workspace = true }
-ev-storage-proofs = { workspace = true }
-hex = { workspace = true }
-serde = { workspace = true }
-sha3 = { workspace = true }
