+++ crates/sp1/ev-hyperlane/program/Cargo.toml
@@ -5,5 +5,5 @@ edition = "2021"
 
 [dependencies]
 alloy-primitives = { workspace = true }
-ev-hyperlane-types = { workspace = true }
+ev-zkevm-types = { workspace = true }
 sp1-zkvm = { workspace = true }
