+++ crates/storage/Cargo.toml
@@ -10,8 +10,7 @@ test = []
 anyhow = { workspace = true }
 bincode = { workspace = true }
 dotenvy = "0.15.7"
-ev-hyperlane-types = { workspace = true }
-ev-state-types = { workspace = true }
+ev-zkevm-types = { workspace = true }
 hex = { workspace = true }
 rocksdb = "0.24.0"
 serde = { workspace = true }
