+++ crates/storage/src/hyperlane/mod.rs
@@ -7,8 +7,9 @@ mod tests {
         message::{HyperlaneMessageStore, IndexMode},
         snapshot::HyperlaneSnapshotStore,
     };
-    use ev_hyperlane_types::tree::MerkleTree;
-    use ev_state_types::{StoredHyperlaneMessage, hyperlane::decode_hyperlane_message};
+    use ev_zkevm_types::{
+        StoredHyperlaneMessage, hyperlane::decode_hyperlane_message, programs::hyperlane::tree::MerkleTree,
+    };
 
     #[test]
     fn test_insert_message() {
